+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0
+
+* Update README
+* Fix controller's queue logic  
+* Add loading wave fade animation
+
 ## 0.0.1
 
 * Initial release: Pinpad and PinIndicator (24 animations)